+++ src/Convert/FloatConverter.php
@@ -6,24 +6,29 @@ class FloatConverter implements TypeConverter
 {
     public function fromGraphQL($value): float
     {
-        // JSON floats can appear like int's
-        if (is_int($value)) {
-            $value = (float) $value;
-        }
-
-        if (! is_float($value)) {
-            throw new \InvalidArgumentException('Expected float, got ' . gettype($value));
-        }
-
-        return $value;
+        return $this->toFloat($value);
     }
 
     public function toGraphQL($value): float
     {
-        if (! is_float($value)) {
-            throw new \InvalidArgumentException('Expected float, got ' . gettype($value));
+        return $this->toFloat($value);
+    }
+
+    /**
+     * @param mixed $value Should be float
+     */
+    protected function toFloat($value): float
+    {
+        // JSON floats can appear like ints
+        if (is_int($value)) {
+            $value = (float) $value;
         }
 
-        return $value;
+        if (is_float($value)) {
+            return $value;
+        }
+
+        $notFloat = gettype($value);
+        throw new \InvalidArgumentException("Expected float, got {$notFloat}");
     }
 }