+++ src/Convert/IntConverter.php
@@ -6,17 +6,22 @@ class IntConverter implements TypeConverter
 {
     public function fromGraphQL($value): int
     {
-        if (! is_int($value)) {
-            throw new \InvalidArgumentException('Expected int, got ' . gettype($value));
-        }
-
-        return $value;
+        return $this->toInt($value);
     }
 
     public function toGraphQL($value): int
     {
+        return $this->toInt($value);
+    }
+
+    /**
+     * @param mixed $value Should be int
+     */
+    protected function toInt($value): int
+    {
         if (! is_int($value)) {
-            throw new \InvalidArgumentException('Expected int, got ' . gettype($value));
+            $notInt = gettype($value);
+            throw new \InvalidArgumentException("Expected int, got {$notInt}");
         }
 
         return $value;