+++ src/Convert/ListConverter.php
@@ -17,7 +17,8 @@ class ListConverter implements TypeConverter
     public function fromGraphQL($value): array
     {
         if (! is_array($value)) {
-            throw new \InvalidArgumentException('Expected array, got ' . gettype($value));
+            $notArray = gettype($value);
+            throw new \InvalidArgumentException("Expected array, got {$notArray}");
         }
 
         // @phpstan-ignore-next-line Parameter #1 $callback of function array_map expects (callable(mixed): mixed)|null, array{Spawnia\Sailor\TypeConverter, 'fromGraphQL'} given.
@@ -30,7 +31,8 @@ class ListConverter implements TypeConverter
     public function toGraphQL($value): array
     {
         if (! is_array($value)) {
-            throw new \InvalidArgumentException('Expected array, got ' . gettype($value));
+            $notArray = gettype($value);
+            throw new \InvalidArgumentException("Expected array, got {$notArray}");
         }
 
         $graphQLValues = array_map([$this->ofType, 'toGraphQL'], $value);