+++ src/Type/FloatTypeConfig.php
@@ -4,14 +4,19 @@ namespace Spawnia\Sailor\Type;
 
 use Spawnia\Sailor\Convert\FloatConverter;
 
-class FloatTypeConfig implements TypeConfig
+class FloatTypeConfig implements TypeConfig, InputTypeConfig, OutputTypeConfig
 {
     public function typeConverter(): string
     {
         return FloatConverter::class;
     }
 
-    public function typeReference(): string
+    public function inputTypeReference(): string
+    {
+        return 'float|int';
+    }
+
+    public function outputTypeReference(): string
     {
         return 'float';
     }